insert into payments and chargehistory
I am having trouble with customers that do not have an account. I tried
to disable the account button however was unsuccessful. Also customers
are able to submit credit card charges without entering a credit card. I
couldnt figure out how to stop the intial enter of the shipment if this
wasnt filled in.